refactor and move stub mojang impl into lib to add unit test in profile_service
