+++ src/lib.rs
@@ -2,5 +2,5 @@ pub mod cache;
 pub mod error;
 pub mod metrics_server;
 pub mod mojang;
-pub mod service;
+pub mod profile_service;
 pub mod settings;
